Extract Payjoin wallet actor into its module
Keep the session lifecycle, broadcast helpers, and
ActivePayjoin state together so actor.rs and
transactions.rs stay focused on their primary concerns.
